move all Unit tests to the Unit namespace
(cherry picked from commit ea507673ee0d72b24c66a1dd2ed363d5593ba5b6)
